backend/src/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
atlassian-python-api==3.41.0
PyGithub==1.59.1
requests==2.31.0
python-multipart==0.0.6
jinja2==3.1.2
reportlab==4.0.7
matplotlib==3.8.2
pandas==2.1.4
asyncio==3.4.3
aiofiles==23.2.1
